// schemas/Contact/contact.schemas.ts
import { z } from "zod"

export const contactSchema = z.object({
  name: z.string().min(2, "Please enter your full name"),

  email: z.email("Please enter a valid email address"),

  subject: z.string().min(3, "Please enter a subject"),

  message: z.string().min(10, "Message must be at least 10 characters"),
})

export type ContactFormValues = z.infer<typeof contactSchema>